requirements.txt
pymysql
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn

